feat: add basic audio support with speech-to-text and text-to-speech
- Add audio_service with Whisper transcription and OpenAI TTS synthesis
- Extract chat logic into chat_service for reuse by voice handlers
- Add voice_message_request and text_to_speech_request websocket handlers
- Add VoiceMessageRequestDto, VoiceTranscribedDto, and TTS DTOs
- Add message_id field to track individual messages across requests
